novos ajustes realizados
