+++ src/main/java/com/contapagar/domain/model/Conta.java
@@ -12,7 +12,7 @@ public class Conta {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 255, columnDefinition = "TEXT")
+    @Column(nullable = false, length = 255)
     private String descricao;
 
     private BigDecimal valor;
