+++ src/main/resources/application.yml
@@ -4,6 +4,7 @@ spring:
     username: postgres
     password: postgres
   jpa:
+    database-platform=org: org.hibernate.dialect.PostgreSQLDialect
     hibernate:
       ddl-auto: update
     show-sql: true
@@ -12,4 +13,4 @@ spring:
     locations: classpath:db/migration
 
 server:
-  port: 8080
+  port: 8081
